+++ src/app/Components/ComponentsSamples/Searching/NativeSearchPageSamples.xaml.cs
@@ -1,0 +1,9 @@
+namespace Components.ComponentsSamples.Searching;
+
+public partial class NativeSearchPageSamples
+{
+    public NativeSearchPageSamples()
+    {
+        InitializeComponent();
+    }
+}