+++ src/library/DIPS.Mobile.UI/Components/Pages/Search/dotnet/SearchBehavior.cs
@@ -1,0 +1,12 @@
+namespace DIPS.Mobile.UI.Components.Pages.Search;
+
+public partial class SearchBehavior
+{
+    partial void SetupNativeSearch(ContentPage page)
+    {
+    }
+
+    partial void TeardownNativeSearch(ContentPage page)
+    {
+    }
+}